Move `reset_history` to `indexes.db`.
