+++ virtool/indexes/db.py
@@ -10,6 +10,7 @@ import pymongo
 import virtool.api.utils
 import virtool.history.db
 import virtool.utils
+from virtool.db.core import DB
 
 PROJECTION = [
     "_id",
@@ -226,3 +227,26 @@ async def get_unbuilt_stats(db, ref_id: Union[str, None] = None) -> dict:
         "change_count": await db.history.count_documents(history_query),
         "modified_otu_count": len(await db.history.distinct("otu.id", history_query))
     }
+
+
+async def reset_history(db: DB, index_id: str):
+    """
+    Set the index.id and index.version fields with the given index id to 'unbuilt'.
+
+    :param db: The virtool database
+    :param index_id: The ID of the index which failed to build
+    """
+    query = {
+        "_id": {
+            "$in": await db.history.distinct("_id", {"index.id": index_id})
+        }
+    }
+
+    return await db.history.update_many(query, {
+        "$set": {
+            "index": {
+                "id": "unbuilt",
+                "version": "unbuilt"
+            }
+        }
+    })
